requirements.txt
pandas
numpy
matplotlib
seaborn
dash
dash-bootstrap-components
plotly-express
scikit-learn
gunicorn